runtime: move genesis_parse to correct dir
