+++ src/discof/genesis/fd_genesi_tile.c
@@ -5,7 +5,7 @@
 #include "../../disco/topo/fd_topo.h"
 #include "../../discof/fd_accdb_topo.h"
 #include "../../ballet/sha256/fd_sha256.h"
-#include "../../flamenco/runtime/fd_genesis_parse.h"
+#include "../../flamenco/genesis/fd_genesis_parse.h"
 #include "../../flamenco/accdb/fd_accdb_admin_v1.h"
 #include "../../flamenco/accdb/fd_accdb_admin_v2.h"
 #include "../../flamenco/accdb/fd_accdb_sync.h"
