+++ src/discof/genesis/fd_genesi_tile.h
@@ -7,7 +7,7 @@
    blob. */
 
 #include "../../ballet/lthash/fd_lthash.h"
-#include "../../flamenco/runtime/fd_genesis_parse.h"
+#include "../../flamenco/genesis/fd_genesis_parse.h"
 
 #define FD_GENESIS_TILE_MTU (sizeof(fd_genesis_meta_t) + FD_GENESIS_MAX_MESSAGE_SIZE)
