+++ src/flamenco/genesis/fd_genesis_parse.c
@@ -1,5 +1,5 @@
 #include "fd_genesis_parse.h"
-#include "fd_runtime_const.h"
+#include "../runtime/fd_runtime_const.h"
 #include "../../util/bits/fd_bits.h"
 
 fd_genesis_t *
