+++ src/flamenco/genesis/fuzz_genesis_parse.c
@@ -1,5 +1,5 @@
 #include "fd_genesis_parse.h"
-#include "fd_runtime_const.h"
+#include "../runtime/fd_runtime_const.h"
 #include <assert.h>
 #include <stdlib.h>
